backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
passlib==1.7.4
numpy
pandas
scikit-learn==1.5.2
bcrypt==4.0.1
gunicorn==21.2.0


google-generativeai
